code-py/requirements-org.txt
azure-core==1.29.6
azure-identity==1.15.0
docx2txt==0.8
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.12
langchain-openai==0.0.2.post1
msgraph-core==0.2.2
openai==1.7.2
tiktoken==0.5.2
